+++ wowp/guild.go
@@ -37,7 +37,7 @@ type Guild struct {
 					R int     `json:"r"`
 					G int     `json:"g"`
 					B int     `json:"b"`
-					A float64 `json:"a"`
+					A float32 `json:"a"`
 				} `json:"rgba"`
 			} `json:"color"`
 		} `json:"emblem"`
@@ -54,7 +54,7 @@ type Guild struct {
 					R int     `json:"r"`
 					G int     `json:"g"`
 					B int     `json:"b"`
-					A float64 `json:"a"`
+					A float32 `json:"a"`
 				} `json:"rgba"`
 			} `json:"color"`
 		} `json:"border"`
@@ -65,7 +65,7 @@ type Guild struct {
 					R int     `json:"r"`
 					G int     `json:"g"`
 					B int     `json:"b"`
-					A float64 `json:"a"`
+					A float32 `json:"a"`
 				} `json:"rgba"`
 			} `json:"color"`
 		} `json:"background"`
